add footer, refactor components to components folder
