+++ src/App.vue
@@ -1,11 +1,13 @@
 <script setup lang="ts">
 import MainHeader from "./components/MainHeader.vue";
-import ContentContainer from "./ContentContainer.vue";
+import ContentContainer from "./components/ContentContainer.vue";
+import MainFooter from "./components/MainFooter.vue";
 </script>
 
 <template>
   <div class="bg-bg min-h-screen text-text-light">
     <MainHeader />
     <ContentContainer />
+    <MainFooter />
   </div>
 </template>
