+++ src/components/MainFooter.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts"></script>
+
+<template>
+  <footer
+    class="h-64 flex justify-center bg-gradient-to-b from-transparent to-black"
+  >
+    <a href="#" class="mt-32 hover:text-text-light-hover text-glow"
+      >Back to top</a
+    >
+  </footer>
+</template>
